Add a Set-StepPending function to Assertions
In Cucumber, when you write a Feature file and then run cucmuber on it
without implementing any steps whatsoever, the result is that the
scenario (and steps) are considered 'undefined'. For Pester, we
currently treat these as Inconclusive (which is fine--we can synonomize
Inconclusive with Undefined). However, there's no explicit mechanism to
mark a step as pending.

This commit adds a Set-StepPending "assertion" which is similar in
nature to the Set-TestInconclusive "assertion". This will mark that step
as 'Pending'.

This commit only adds the Set-StepPending function. The next series of
commits will update Pester's Gherkin implementation to use it.